fix: address PO review — split WaveguideRouter past size limit, update nazca round-trip expectations for curved metal
- Split path-blocking checks into WaveguideRouter.Blocking.cs partial (512 → 425 lines, under the 500-line hard limit).
- Update GdsMziElectricalRoundTripTests for intended curved-metal geometry (#854): metal traces now bend, so layer=(11,0) statements 31→38 and flattened metal polygons 29→36; detector_cross traces no longer cross, so ALL FOUR metal traces restore as electrical connections (was 2 + a metal junction); the wider curved traces push two optical routes into a second optical junction (optical restored 4→2, frozen polys 24→22).
- Verified pin anchors stay put: anode OffsetX unchanged, and new assertions pin the anode↔cathode 55 µm template spacing — only the bbox re-frame moved OffsetY (225.20→215.94).
- Both nazca-gated tests run and pass locally with real nazca.
